further amends
Phil and I talked about it this morning, and we realized we don’t really
need a static map anymore since the PPT essentially covers that (and
anyone could save it as a PDF if they wanted to). Our long term plan is
to add competitor maps, so could we add a coming soon button for those
instead (to the right of the Interactive Map and PowerPoint Summaries
buttons if that’s possible)?

For the PowerPoint page, could we do the description and then a large
image of slide 1 (with an embedded link)? Let me know if you had
something else in mind. Here are the descriptions:

Interactive Map description: This interactive map allows you to click
through each stage of the author journey and see the various services
Wiley offers researchers, including a brief description and a link to
more information (where available). Each service is color-coded based on
whether it is a tool or product, a learning service, or a policy-based
service.

PowerPoint Summary description: This PowerPoint is a static version of
the interactive map that can be used in presentations or saved as a PDF
and shared. The first slide shows all of Wiley’s services across the
stages of the research journey, and the subsequent slides isolate each
stage. Links to the external webpage for each service are included where
available.

Change to the “About” section on the main page (in red): These author
services maps show the various services that Wiley offers Researchers
across the research journey – from discovery through to assessment.
There is an Interactive Map which allows you to click through each stage
of the author journey and see the services, including a brief
description. The PowerPoint Summary shows an overview of the entire
author journey and the services offered (with embedded links) and is
available to use for presentations or to save as a PDF. Click on the
relevant graphic for more information. The Competitor Maps page (coming
soon) will show what services other publishers are currently offering.